clean up genenv, add to startserver, update dockerfile with genenv

+++ genenv.py
@@ -1,10 +1,16 @@
 #!/usr/bin/env python3
 
+from os.path import isfile
 from django.core.management import utils
 
-new_secret = utils.get_random_secret_key()
+# Bail if .env exists already
+if isfile('.env'):
+    print('.env file already present; exiting.')
+    exit(0)
+
 
 try:
+    new_secret = utils.get_random_secret_key()
     with open('example.env','rt') as env_in:
         with open('.env', 'wt') as env_out:
             for line in env_in:
